refactor(gui): split records list modules
Refs #66
